+++ scripts/util.py
@@ -415,6 +415,8 @@ def load_config_data(config):
         cfg.slurm_bin_dir = slurmdirs.prefix / "bin"
     if not cfg.slurm_control_host:
         cfg.slurm_control_host = f"{cfg.slurm_cluster_name}-controller"
+    if not cfg.slurm_control_host_port:
+        cfg.slurm_control_host_port = "6820-6830"
 
     if not cfg.enable_debug_logging and isinstance(cfg.enable_debug_logging, NSDict):
         cfg.enable_debug_logging = False
@@ -454,7 +456,7 @@ def config_from_metadata():
     for retry, wait in enumerate(backoff_delay(0.25, count=4, timeout=12)):
         config_yaml = project_metadata.__wrapped__(metadata_key)
         if config_yaml is None:
-            log.error(f"config not found in project metadata, retry {retry}")
+            log.warning(f"config not found in project metadata, retry {retry}")
             sleep(wait)
             continue
         else:
@@ -471,7 +473,7 @@ def load_config_file(path):
     try:
         content = yaml.safe_load(Path(path).read_text())
     except FileNotFoundError:
-        log.error(f"config file not found: {path}")
+        log.warning(f"config file not found: {path}")
         return NSDict()
     return load_config_data(content)
 
@@ -1089,8 +1091,16 @@ class Lookup:
         return self.cfg.project or auth_project
 
     @property
+    def control_addr(self):
+        return self.cfg.slurm_control_addr
+
+    @property
     def control_host(self):
         return self.cfg.slurm_control_host
+
+    @property
+    def control_host_port(self):
+        return self.cfg.slurm_control_host_port
 
     @property
     def scontrol(self):
@@ -1434,7 +1444,6 @@ class Lookup:
 # Define late globals
 cfg = load_config_file(CONFIG_FILE)
 if not cfg:
-    log.warning(f"{CONFIG_FILE} not found")
     cfg = config_from_metadata()
     if cfg:
         save_config(cfg, CONFIG_FILE)